applied nz modal for udating profile

+++ src/app/dashbord/user-list/user-list.component.ts
@@ -2,9 +2,7 @@ import { Component } from '@angular/core';
 import { Router } from '@angular/router';
 import { ToastrService } from 'ngx-toastr';
 import { UserdataService } from 'src/app/service/userdata.service';
-import { getAuth, deleteUser } from 'firebase/auth';
-
-const auth = getAuth();
+import { NzModalService } from 'ng-zorro-antd/modal';
 
 @Component({
   selector: 'app-user-list',
@@ -27,7 +25,7 @@ export class UserListComponent {
   select: any;
   constructor(
     private auth: UserdataService,
-
+    private modal: NzModalService,
     private router: Router,
     private toaster: ToastrService
   ) {
@@ -48,16 +46,26 @@ export class UserListComponent {
   showSuccesEdit() {
     this.toaster.success('User information updated successfully', 'Success');
   }
-  editUser(item: any) {
+  showModal(item: any) {
     this.isEdit = true;
-
     this.updateUserDet.userName = item.userName;
     this.updateUserDet.designation = item.designation;
     this.updateUserDet.number = item.number;
     console.log('user', item);
     localStorage.setItem('selected', JSON.stringify(item));
   }
-  updateUser(data: any, item: any) {
+  handleOk(): void {
+    console.log('Button ok clicked!');
+
+    this.isEdit = false;
+  }
+
+  handleCancel(): void {
+    console.log('Button cancel clicked!');
+    this.isEdit = false;
+  }
+
+  updateAdmin(data: any) {
     this.select = localStorage.getItem('selected');
 
     this.select = JSON.parse(this.select);
@@ -81,17 +89,27 @@ export class UserListComponent {
   }
 
   deleteUser(userId: string) {
-    this.auth.deleteUser(userId).subscribe((res) => {
-      this.data = this.data.filter((eachData: any) => {
-        eachData.userId !== userId;
-      });
-      this.auth.allData().subscribe((resp) => {
-        this.user = Object.values(resp);
-        let rep = this.user.filter((u: any) => this.userData.uid !== u.uid);
-        this.user = rep;
-        deleteUser();
-      });
-      this.showSuccess();
+    this.modal.confirm({
+      nzTitle: 'Are you sure delete this user?',
+      nzContent: '<b style="color: red;">Some descriptions</b>',
+      nzOkText: 'Yes',
+      nzOkType: 'primary',
+      nzOkDanger: true,
+      nzOnOk: () => {
+        this.auth.deleteUser(userId).subscribe((res) => {
+          this.data = this.data.filter((eachData: any) => {
+            eachData.userId !== userId;
+          });
+          this.auth.allData().subscribe((resp) => {
+            this.user = Object.values(resp);
+            let rep = this.user.filter((u: any) => this.userData.uid !== u.uid);
+            this.user = rep;
+          });
+          this.showSuccess();
+        });
+      },
+      nzCancelText: 'No',
+      nzOnCancel: () => console.log('Cancel'),
     });
   }
 }
